refactor: move ContentLength to source module and simplify storage API
- Move ContentLength and DynamicLength from storage to source module
- Simplify StorageProvider::into_reader_writer to accept Option<u64>
- Update StreamDownload::content_length to return Option<u64>
- Add helper constructors for ContentLength variants
- Implement conversions and comparisons between ContentLength and u64
- Update all storage implementations to use simplified signature
