+++ crates/stream-download-opendal/src/lib.rs
@@ -13,8 +13,8 @@ use bytes::{Bytes, BytesMut};
 use futures_util::{Stream, ready};
 use opendal::{FuturesAsyncReader, Operator, Reader};
 use pin_project_lite::pin_project;
-use stream_download::source::{DecodeError, SourceStream};
-use stream_download::storage::{ContentLength, StorageProvider};
+use stream_download::source::{ContentLength, DecodeError, SourceStream};
+use stream_download::storage::StorageProvider;
 use stream_download::{Settings, StreamDownload, StreamInitializationError};
 use tokio_util::compat::{Compat, FuturesAsyncReadCompatExt};
 use tokio_util::io::poll_read_buf;
