+++ crates/stream-download/src/storage/temp.rs
@@ -13,7 +13,7 @@ use educe::Educe;
 pub use tempfile;
 use tempfile::NamedTempFile;
 
-use super::{ContentLength, StorageProvider};
+use super::StorageProvider;
 use crate::WrapIoResult;
 
 type TempfileFnType = Arc<dyn Fn() -> io::Result<NamedTempFile> + Send + Sync + 'static>;
@@ -106,7 +106,7 @@ impl StorageProvider for TempStorageProvider {
 
     fn into_reader_writer(
         self,
-        _content_length: ContentLength,
+        _content_length: Option<u64>,
     ) -> io::Result<(Self::Reader, Self::Writer)> {
         let tempfile = if let Some(tempfile_fn) = self.tempfile_fn {
             (tempfile_fn.0)()
